Finish off all integration tests for v31

+++ integration_test/tests/hidden.rs
@@ -2,10 +2,6 @@
 
 //! Tests for methods that are `== Hidden ==` and not in the API docs of Bitcoin Core.
 
-// Hidden RPCs are absent from Core's OpenRPC spec, so they have no generated types or
-// `into_model`; with the curated conversions retired these tests only exist for the sync
-// client, which is itself disabled in the async migration.
-#![cfg(not(feature = "test-async"))]
 #![allow(non_snake_case)] // Test names intentionally use double underscore.
 
 #[cfg(not(feature = "v28_and_below"))]
@@ -288,7 +284,7 @@ fn hidden__get_raw_addrman() {
     let entry = json
         .new
         .values()
-        .find(|e| e.address == peer_address && e.port == peer_port)
+        .find(|e| e.address == peer_address && e.port as i64 == peer_port as i64)
         .expect("added peer should appear in the 'new' table");
 
     assert_eq!(entry.network, "ipv4");
